+++ test/fixtures/backfill_worker.rb
@@ -1,0 +1,11 @@
+class BackfillWorker
+  include Sidekiq::Worker
+  include Sidetiq::Schedulable
+
+  tiq backfill: true do
+    daily
+  end
+
+  def perform
+  end
+end