+++ test/test_schedule.rb
@@ -17,5 +17,22 @@ class TestSchedule < Sidetiq::TestCase
     assert sched.schedule_next?(Time.now + (2 * 24 * 60 * 60))
     refute sched.schedule_next?(Time.now + (2 * 24 * 60 * 60))
   end
+
+  def test_backfill
+    sched = Sidetiq::Schedule.new
+    refute sched.backfill?
+    sched.backfill = true
+    assert sched.backfill?
+  end
+
+  def test_set_options
+    sched = Sidetiq::Schedule.new
+
+    sched.set_options(backfill: true)
+    assert sched.backfill?
+
+    sched.set_options(backfill: false)
+    refute sched.backfill?
+  end
 end